requirements.txt
anthropic>=0.40.0
faster-whisper>=1.0.0
python-docx>=1.1.0
numpy>=1.26.0
streamlit>=1.35.0
audio-recorder-streamlit>=0.0.8
